Fix compilation warning/error.
On OSX 10.7 after updating to ruby 2 I was getting this warning
turned error:

boolarray.h:137: warning: assuming signed overflow does not occur
when simplifying multiplication

I fixed it by casting the return value of buffer.size() to uint.

+++ ext/boolarray.h
@@ -134,7 +134,7 @@ public:
      * set all bits to 0
      */
     void reset() {
-        memset(&buffer[0],0,sizeof(uword)*buffer.size());
+        memset(&buffer[0],0,sizeof(uword)*(uint)buffer.size());
         sizeinbits = 0;
     }
